requirements.txt
Django==1.10.3
pushbullet.py==0.10.0
requests==2.11.1